# frontend/moderator-dashboard-ui/docs/verification/backend/openapi_verified.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B2B Platform API",
    "description": "API for B2B Platform - supplier moderation and parsing system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/moderator/suppliers/{supplier_id}/keywords": {
      "get": {
        "tags": [
          "Suppliers"
        ],
        "summary": "Get Supplier Keywords Endpoint",
        "description": "Get supplier keywords.",
        "operationId": "get_supplier_keywords_endpoint_moderator_suppliers__supplier_id__keywords_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Supplier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierKeywordsResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/keywords": {
      "get": {
        "tags": [
          "Keywords"
        ],
        "summary": "List Keywords Endpoint",
        "description": "List all keywords.",
        "operationId": "list_keywords_endpoint_keywords_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordsListResponseDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Keywords"
        ],
        "summary": "Create Keyword Endpoint",
        "description": "Create a new keyword.",
        "operationId": "create_keyword_endpoint_keywords_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeywordRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/keywords/{keyword_id}": {
      "delete": {
        "tags": [
          "Keywords"
        ],
        "summary": "Delete Keyword Endpoint",
        "description": "Delete keyword.",
        "operationId": "delete_keyword_endpoint_keywords__keyword_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Keyword Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/learning/learn-manual-inn": {
      "post": {
        "tags": [
          "Learning"
        ],
        "summary": "Learn Manual Inn",
        "description": "Ручное обучение Domain Parser по ссылке, где найден ИНН.",
        "operationId": "learn_manual_inn_learning_learn_manual_inn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnManualInnRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnManualInnResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/learning/statistics": {
      "get": {
        "tags": [
          "Learning"
        ],
        "summary": "Get Learning Statistics",
        "description": "Получить статистику обучения Domain Parser.",
        "operationId": "get_learning_statistics_learning_statistics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningStatisticsDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/learning/learned-summary": {
      "get": {
        "tags": [
          "Learning"
        ],
        "summary": "Get Learned Summary",
        "description": "Получить краткую сводку выученных паттернов.",
        "operationId": "get_learned_summary_learning_learned_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateKeywordRequestDTO": {
        "properties": {
          "keyword": {
            "type": "string",
            "title": "Keyword"
          }
        },
        "type": "object",
        "required": [
          "keyword"
        ],
        "title": "CreateKeywordRequestDTO",
        "description": "Request DTO for creating keyword."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "keyword",
          "created_at"
        ],
        "title": "KeywordDTO",
        "description": "Keyword DTO."
      },
      "KeywordsListResponseDTO": {
        "properties": {
          "keywords": {
            "items": {
              "$ref": "#/components/schemas/KeywordDTO"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "keywords"
        ],
        "title": "KeywordsListResponseDTO",
        "description": "Response DTO for keywords list."
      },
      "LearnManualInnRequestDTO": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid",
            "description": "ID parsing run"
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Домен для обучения"
          },
          "inn": {
            "type": "string",
            "title": "Inn",
            "description": "ИНН, найденный вручную"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl",
            "description": "URL страницы, где найден ИНН"
          },
          "learningSessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learningsessionid",
            "description": "ID сессии обучения"
          }
        },
        "type": "object",
        "required": [
          "runId",
          "domain",
          "inn",
          "sourceUrl"
        ],
        "title": "LearnManualInnRequestDTO",
        "description": "Запрос на ручное обучение по ссылке с ИНН."
      },
      "LearnManualInnResponseDTO": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid",
            "description": "ID parsing run"
          },
          "learningSessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learningsessionid",
            "description": "ID сессии обучения"
          },
          "learnedItems": {
            "items": {
              "$ref": "#/components/schemas/LearnedItemDTO"
            },
            "type": "array",
            "title": "Learneditems",
            "description": "Выученные элементы"
          },
          "statistics": {
            "$ref": "#/components/schemas/LearningStatisticsDTO",
            "description": "Статистика обучения"
          }
        },
        "type": "object",
        "required": [
          "runId",
          "statistics"
        ],
        "title": "LearnManualInnResponseDTO",
        "description": "Ответ на ручное обучение по ИНН."
      },
      "LearnedItemDTO": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain",
            "description": "Домен, на котором обучились"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Тип данных: 'inn' или 'email'"
          },
          "value": {
            "type": "string",
            "title": "Value",
            "description": "Найденное значение (ИНН или Email)"
          },
          "sourceUrls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sourceurls",
            "description": "URL источников, где найдены данные"
          },
          "urlPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Urlpatterns",
            "description": "Выученные URL паттерны"
          },
          "learning": {
            "type": "string",
            "title": "Learning",
            "description": "Описание того, чему научились"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "type",
          "value",
          "learning"
        ],
        "title": "LearnedItemDTO",
        "description": "Информация о выученном элементе."
      },
      "LearningStatisticsDTO": {
        "properties": {
          "totalLearned": {
            "type": "integer",
            "title": "Totallearned",
            "description": "Всего выучено паттернов",
            "default": 0
          },
          "successRateBefore": {
            "type": "number",
            "title": "Successratebefore",
            "description": "Процент успеха до обучения",
            "default": 0.0
          },
          "successRateAfter": {
            "type": "number",
            "title": "Successrateafter",
            "description": "Процент успеха после обучения",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "LearningStatisticsDTO",
        "description": "Статистика обучения."
      },
...
      "SupplierKeywordsResponseDTO": {
        "properties": {
          "keywords": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "keywords"
        ],
        "title": "SupplierKeywordsResponseDTO",
        "description": "Response DTO for supplier keywords."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}